requirements.txt
Django==2.0.3
django-crispy-forms==1.7.2
django-filter==1.1.0
django-tables2==1.21.1
pytz==2018.3
